services/moderation_service/requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14.0
ultralytics>=8.0.0
Pillow>=10.0.0
